remove duplicate functionality across querying and toggling
