+++ src/composables/useTasksVariantsToggleRegistered.js
@@ -1,6 +1,6 @@
-import { ref, watch } from 'vue';
-import useToggleRegisteredTasksMutation from '@/composables/mutations/useToggleRegisteredTasksMutation';
-import useToggleRegisteredVariantsMutation from '@/composables/mutations/useToggleRegisteredVariantsMutation';
+import { ref } from 'vue';
+import useTasksQuery from '@/composables/queries/useTasksQuery';
+import useTaskVariantsQuery from '@/composables/queries/useTaskVariantsQuery';
 
 /**
  * Composable for managing registration state of tasks and variants
@@ -12,25 +12,17 @@ export function useTasksVariantsToggleRegistered() {
   const registeredTasksOnly = ref(localStorage.getItem('registeredTasksOnly') === 'false' ? false : true);
   const registeredVariantsOnly = ref(localStorage.getItem('registeredVariantsOnly') === 'false' ? false : true);
 
-  watch(registeredTasksOnly, (value) => {
-    localStorage.setItem('registeredTasksOnly', value);
-  });
-
-  watch(registeredVariantsOnly, (value) => {
-    localStorage.setItem('registeredVariantsOnly', value);
-  });
-
-  const { mutate: toggleRegisteredTasks } = useToggleRegisteredTasksMutation();
-  const { mutate: toggleRegisteredVariants } = useToggleRegisteredVariantsMutation();
+  const { refetch: toggleRegisteredTasks } = useTasksQuery(registeredTasksOnly.value);
+  const { refetch: toggleRegisteredVariants } = useTaskVariantsQuery(registeredVariantsOnly.value);
 
   const updateRegisteredTasksOnly = (value) => {
     registeredTasksOnly.value = value;
-    toggleRegisteredTasks(value, null);
+    toggleRegisteredTasks();
   };
 
   const updateRegisteredVariantsOnly = (value) => {
     registeredVariantsOnly.value = value;
-    toggleRegisteredVariants(value);
+    toggleRegisteredVariants();
   };
 
   return {
